requirements.txt
Flask==3.0.3
Flask_Cors==4.0.0
imgaug==0.4.0
matplotlib==3.8.3
natsort==8.4.0
numpy==1.25.2
opencv_python==4.9.0.80
opencv_python_headless==4.9.0.80
Pillow==10.3.0
skimage==0.0
torch==2.3.0+cu121
torchvision==0.18.0+cu121
tqdm==4.66.2
ultralytics==8.1.47
waitress==3.0.0
